Fixed bug

+++ src/GameHost.java
@@ -253,7 +253,7 @@ public class GameHost {
 		System.out.println("MOVE MADE IS: " + m);
 
 		String[] data = m.split("\\s+");
-		Move mv = new Move(gameBoard.getPiece(Integer.parseInt(data[0]), Integer.parseInt(data[1])), Integer.parseInt(data[2]), Integer.parseInt(data[3]), Double.parseDouble(data[4]));
+		Move mv = new Move(gameBoard.getPiece(Integer.parseInt(data[0]), Integer.parseInt(data[1])), Integer.parseInt(data[2]), Integer.parseInt(data[3]), Long.parseLong(data[4]));
 		if (SpeedChess.playerPerspective == 0) {
 			players[1].setNextMove(mv);
 		} else {
